refactor: split progress fact applier
